move remaining code out of listener.py
